Process Each File

Process each file according to the specification in Access database.

Usage

do_make_file_each(
  spec,
  fgspec,
  aggregate,
  datacols,
  year,
  row,
  base,
  duck = NULL
)

Arguments

spec

File specification

fgspec

Filegroup specification

aggregate

Logical value. Default is TRUE. Aggregate data according to the specification in registration database. Global options with orgdata.aggregate.

datacols

Columnames to be kept

year

Which reference year to use for geograhical coding. If it is missing then global option for orgdata.year will be used.

row

Select specific row(s) numbers only. Useful for debugging. Please read Debugging article for detail.

base

Logical value. If TRUE then use year in the original data as the base year to recode the geographical codes. Default is FALSE and use all available codes in geo codebook

duck

R6 object for DuckDB