Special Need Reshape to Long

This function is only applicable to reshape data that was reshaped to wide via Access specification in RESHAPE columns.

Usage

do_reshape_long(dt = NULL, respec = NULL)

Arguments

dt

An output dataset from do_reshape_wide()

respec

Reshape specification with id and value variables. This is the output from get_reshape_wide_spec()

See also