Columns with Multiple Inputs

Get all arguments in a selected column that have multiple arguments with find_column_multi(). The output will be a character type of arguments length. This can then can be used in find_column_multi_input() function to get the value of the input as a list object. If you are only interested in a specific argument among these arguments in the column, then use find_column_multi_input_arg() function. See example.

find_column_multi(df = NULL, col = NULL, sep = c(",", ":", ";"))

find_column_multi_input(input = NULL)

find_column_multi_input_arg(input = NULL, arg = NULL)

Arguments

df
Data.frame extracted form registration database

col
Column name in the database table

sep
Symbols that separate these arguments eg. "," or ":"

input
Input argument(s) as a character vector

arg
Name of a specific argument in the column

Value

Two different output:

find_column_multi gives a character vector of the arguments that is separated with sep argument

find_column_multi_input gives a list of argument names and their values

The single object with value from the selected argument

Examples

if (FALSE) {
args <- find_column_multi(spec$INNLESARG)
vals <- find_column_multi_input(args)
val <- find_column_multi_input_arg(args, "header")
}