Geographical Codes to Recode

Create a table of current year geographical codes against previous years geogprahical codes. This is used to recode the previous years codes to the current codes. Implementation of this function is base on norgeo::track_change() function.

Usage

geo_recode(
  type = c("grunnkrets", "bydel", "kommune", "fylke"),
  from = NULL,
  to = NULL,
  write = FALSE,
  append = FALSE,
  fix = TRUE
)

Arguments

type

Type of regional granularity ie. enumeration area (grunnkrets)

from

Starting year for the range period. Current year is the default if left empty

to

End of year for the range period. Current year is the default if left empty

write

Write table to the orgdata.geo database. It will overwrite the table if it already exists

append

Append the data to an existing table in the orgdata.geo

fix

Default is TRUE. Use external codes to fix geo changes manually. The codes is sourced from config files depending on the granularity levels.

See also

Examples

if (FALSE) { # \dontrun{
geo_recode(type = "grunnkrets", from = 2018, to = 2021, write = TRUE)
geo_recode(type = "grunnkrets", from = 2018, to = 2021, append = TRUE)
} # }