Create Dummy Enumeration Area Codes

Some of the downloaded enumeration area codes from SSB lack codes for missing ie. xxxx9999. This function create these codes that are needed for recoding older codes to current enumeration area codes.

get_geo_dummy(dt, from, to)

Arguments

dt

Downloaded data with norgeo::track_change()

from

Starting year for the range period. Current year is the default if left empty

to

End of year for the range period. Current year is the default if left empty

See also