Read Data File

Read rawdata either using FILID value or complete file path. It uses the find_data() generic method. For a .csv file, data.table::fread() is used and all other arguments for fread function can be used. For a .xlsx or .xls file readxl::read_excel() function and all of its arguments.

Nevertheless, some most used arguments are standardized for read_file() and there are:

nrows to display maksimum numbers to read

header FALSE to give default columnames as V1, V2 etc

skip a specific number of raws before reading the data

trimws to trim leading and trailing whitespace

na for character value to be interpreted as NA

read_file(file = NULL, ...)

les_fil(file = NULL, ...)

Arguments

file
Use FILID or a complete path of a filename

...
All other arguments to be passed related to the file format

Examples

if (FALSE) {
# With FILID
DT <- read_file(file = 5)
DT <- read_file(file = 5, fill = TRUE, nrows = 10)

# With filepath
rdata <- read_file(file = "/file/path/mydata.xlsx", sheet = "S3", range = cell_rows(1:4))
rdata <- read_file(file = "/file/path/mydata.csv", sep = ",", header = FALSE)
}