Access Structured Data in Data Warehouse

See or delete the original data in data warehouse that have been cleaned and structured. Data is saved in the warehouse when column KONTROLLERT is marked in the original file table in Access database. This indicates that dataset has been checked and recoded as specified in INNLESING table in Access registration database.

Usage

see_data(
  group = NULL,
  koblid = NULL,
  year = NULL,
  action = c("read", "delete")
)

Arguments

group

The filegroup name (filgruppe)

koblid

Specify one or multiple KOBLID. Use "all" to delete all data in warehouse for seleted filegroup.

year

Which reference year to use for geograhical coding. If it is missing then global option for orgdata.year will be used.

action

To read or delete the data in the warehouse. Default is read.

Examples

if (FALSE) { # \dontrun{
dt <- see_data("LESEFERD", koblid = 134)
dt <- see_data("ENPERSON", koblid = "all", action = "delete")
dt <- see_data("ENPERSON", koblid = 267:270, action = "delete")
} # }