File Specifications in Registration Database

This function will find all specifications in the Access registration database via SQL code. For example where the orginal data is and how the data will be restructured and aggregated.. etc.. etc.. The specifications are registered in the following tables:

tbl_Filgruppe

tbl_Koble

tbl_Orgfile

tbl_Innlesing

SQL file must be written with base::sprintf style ie. '%s','%d' etc. Please refer to base::sprintf documentation. In addition, SQL code must not contain comments. Example of SQL code with '%s': SELECT * FROM tbl_Koble WHERE FILGRUPPE = '%s' Which is saved in C:/myfile.sql and run as in the code example.

find_spec(
  file = NULL,
  value = NULL,
  con = NULL,
  external = FALSE,
  char = NULL,
  num = NULL,
  opposite = FALSE
)

find_query(
  file = NULL,
  value = NULL,
  external = FALSE,
  char = NULL,
  num = NULL,
  opposite = FALSE
)

Arguments

file
SQL file. If external is TRUE then complete filepath must be specified.

value
The value for selection in SQL code with base::sprintf style. For example the name of a filgruppe.

con
Connection to database

external
If the SQL file is outside of the package. Default is FALSE.

char
Which value are character

num
Which value are numeric

opposite
TRUE if numeric value will be read before character

Value

Out put will be a data.frame.

Examples

if (FALSE) {
qr <- find_query("C:/myfile.sql", value = "BEFOLKNING", external = TRUE)
qr2 <- find_query("your.sql", char = "BEFOLKNING", num = 14)
}