Save Data to CSV file

Save data as a .csv format file with semicolon ; as seperation. The file will be saved to the specified folder as in getOption("orgdata.folder.output"). Use argument save = TRUE in read_raw() will activate save_file() directly. Else you can call save_file() to save the object output from read_raw()

save_file(dt = NULL, group = NULL, date = FALSE)

lagrefil(dt = NULL, group = NULL, date = FALSE)

Arguments

dt
Input data preferrably in a data.table format but is not mandatory

group
The group of files (filgruppe)

date
Output file will be named with date and time

Examples

if (FALSE) {
 opitions(orgdata.aggregate = TRUE)
 DF <- read_raw("BEFOLKNING", save = TRUE)

 # Two steps
 DF <- read_raw("BEFOLKNING")
 save_file(DF, "BEFOLKNING")
}